refactor(app): define module boundaries
Restructures the source tree into app, modules, and shared areas.
