+++ src/modules/i18n/index.ts
@@ -1,0 +1,1 @@
+export {};
