+++ src/modules/item/index.ts
@@ -1,0 +1,1 @@
+export {};
