+++ src/modules/reservation/index.ts
@@ -1,0 +1,1 @@
+export {};
