+++ src/modules/share/index.ts
@@ -1,0 +1,1 @@
+export {};
