+++ src/modules/wishlist/index.ts
@@ -1,0 +1,1 @@
+export {};
